test(BA-7779): one test module per behaviour instead of one table per domain
A domain's whole behaviour in a single 357-line table was hard to read and hard to
review. Each behaviour now has its own module holding the rows that describe it, at
95 to 134 lines: creating, reading, searching, editing, retiring.

Passing the rows through a fixture, which was the other idea, cannot work. pytest
parametrizes at collection time and resolves fixtures at setup time, so a fixture has
no way to supply the rows. The two fixture-shaped routes are a params= fixture, which
is still collection-time data, and indirect=True, where the fixture receives a param
rather than providing one.

A scenarios package that the test modules import was the other candidate, and it earns
its keep only when more than one module runs the same rows. With a single runner it
buys nothing but an import list to keep in step, so the rows sit beside the behaviour
they describe. The shared vocabulary moves to the domain's wiring module, which is not
a test file, and the runner to the domain conftest.

Pants now sees five targets where it saw one, so a change reruns the behaviour it
touched and the shards balance better.
