+++ tests/scenario/bai_scenario/manager/domain/conftest.py
@@ -1,4 +1,8 @@
-"""Domain behaviour through the transport-agnostic adapter."""
+"""What every domain scenario module needs: a runner bound to the domain's wiring.
+
+The rows themselves stay in the test modules beside the behaviour they describe. Only
+the runner is shared, which is what a conftest is for.
+"""
 
 from __future__ import annotations
 
@@ -10,7 +14,6 @@ from bai_kit.manager.db import TemplateDatabase
 from bai_kit.manager.monitors import ActionRecorder
 from bai_kit.manager.typed_runner import TypedRunner
 from bai_kit.manager.wiring.domain import domain_wiring
-from bai_scenario.manager.domain.scenarios import SCENARIOS, DomainScenario
 
 
 @pytest.fixture
@@ -27,8 +30,3 @@ def run(
         base_config=base_config_dict(world_template.addr, test_db, None),
         recorder=recorder,
     )
-
-
-@pytest.mark.parametrize("scenario", SCENARIOS, ids=lambda s: s.id)
-async def test_domain(scenario: DomainScenario, run: TypedRunner) -> None:
-    await run(scenario)
