+++ tests/scenario/bai_scenario/manager/domain/test_creating.py
@@ -1,0 +1,116 @@
+"""Creating a domain: who may, and what the new domain comes back as."""
+
+from __future__ import annotations
+
+from datetime import UTC, datetime
+from uuid import UUID
+
+import pytest
+from bai_kit.manager.personas import DOMAIN_ADMIN, MEMBER
+from bai_kit.manager.typed_runner import TypedRunner
+from bai_kit.manager.wiring.domain import (
+    ENFORCEMENT_OFF,
+    WITHIN_THE_RUN,
+    DomainScenario,
+    admin_create,
+    already_there,
+)
+
+from ai.backend.common.data.entity.domain import DomainID
+from ai.backend.common.dto.manager.v2.domain.request import (
+    CreateDomainInput,
+)
+from ai.backend.common.dto.manager.v2.domain.response import (
+    DomainBasicInfo,
+    DomainLifecycleInfo,
+    DomainNode,
+    DomainPayload,
+    DomainRegistryInfo,
+)
+from ai.backend.common.exception import InvalidAPIParameters
+from ai.backend.manager.errors.auth import InsufficientPrivilege
+from ai.backend.testutils.typed_scenario import (
+    TypedScenario,
+    at,
+    checked,
+    exactly,
+    ignored,
+)
+
+SCENARIOS: list[DomainScenario] = [
+    TypedScenario.ok(
+        "superadmin-creates-a-domain-and-gets-the-whole-node-back",
+        when=admin_create(CreateDomainInput(name="new-domain", description="a fresh one")),
+        then=exactly(
+            DomainPayload(
+                domain=DomainNode(
+                    id=DomainID(UUID(int=0)),
+                    basic_info=DomainBasicInfo(
+                        name="new-domain", description="a fresh one", integration_name=None
+                    ),
+                    registry=DomainRegistryInfo(allowed_docker_registries=[]),
+                    lifecycle=DomainLifecycleInfo(
+                        is_active=True,
+                        is_default=False,
+                        created_at=datetime.now(UTC),
+                        modified_at=datetime.now(UTC),
+                    ),
+                )
+            ),
+            where=(
+                ignored(lambda p: p.domain.id, "generated by the database"),
+                checked(lambda p: p.domain.lifecycle.created_at, WITHIN_THE_RUN, "just written"),
+                checked(lambda p: p.domain.lifecycle.modified_at, WITHIN_THE_RUN, "just written"),
+            ),
+        ),
+    ),
+    TypedScenario.ok(
+        "a-domain-is-created-active-unless-the-request-says-otherwise",
+        when=admin_create(CreateDomainInput(name="dormant", is_active=False)),
+        then=at(lambda p: p.domain.lifecycle.is_active, False),
+    ),
+    TypedScenario.ok(
+        "the-allowed-registries-of-the-request-are-what-the-domain-gets",
+        when=admin_create(
+            CreateDomainInput(name="with-registries", allowed_docker_registries=["cr.example.io"])
+        ),
+        then=at(lambda p: p.domain.registry.allowed_docker_registries, ["cr.example.io"]),
+    ),
+    TypedScenario.error(
+        "a-name-already-taken-is-refused",
+        given=[already_there("taken")],
+        when=admin_create(CreateDomainInput(name="taken")),
+        then=InvalidAPIParameters,
+    ),
+    TypedScenario.error(
+        "a-blank-name-is-refused",
+        when=admin_create(CreateDomainInput(name="   ")),
+        then=InvalidAPIParameters,
+    ),
+    TypedScenario.error(
+        "the-domain-admin-may-not-create-a-domain",
+        actor=DOMAIN_ADMIN,
+        when=admin_create(CreateDomainInput(name="by-domain-admin")),
+        then=InsufficientPrivilege,
+    ),
+    TypedScenario.error(
+        "a-member-may-not-create-a-domain",
+        actor=MEMBER,
+        when=admin_create(CreateDomainInput(name="by-member")),
+        then=InsufficientPrivilege,
+    ),
+    TypedScenario.error(
+        # The gate is the role, not the permission graph, so turning enforcement off
+        # changes nothing here. That is the rule this row records.
+        "turning-enforcement-off-still-does-not-let-a-member-create-a-domain",
+        actor=MEMBER,
+        setup=ENFORCEMENT_OFF,
+        when=admin_create(CreateDomainInput(name="by-member-again")),
+        then=InsufficientPrivilege,
+    ),
+]
+
+
+@pytest.mark.parametrize("scenario", SCENARIOS, ids=lambda s: s.id)
+async def test_creating(scenario: DomainScenario, run: TypedRunner) -> None:
+    await run(scenario)
