+++ tests/scenario/bai_scenario/manager/domain/test_editing.py
@@ -1,0 +1,97 @@
+"""Editing a domain, and who may."""
+
+from __future__ import annotations
+
+from datetime import UTC, datetime
+from uuid import UUID
+
+import pytest
+from bai_kit.manager.personas import DOMAIN_ADMIN, MEMBER
+from bai_kit.manager.typed_runner import TypedRunner
+from bai_kit.manager.wiring.domain import (
+    WITHIN_THE_RUN,
+    DomainScenario,
+    admin_update,
+    already_there,
+)
+
+from ai.backend.common.data.entity.domain import DomainID
+from ai.backend.common.dto.manager.v2.domain.request import (
+    UpdateDomainInput,
+)
+from ai.backend.common.dto.manager.v2.domain.response import (
+    DomainBasicInfo,
+    DomainLifecycleInfo,
+    DomainNode,
+    DomainPayload,
+    DomainRegistryInfo,
+)
+from ai.backend.manager.errors.permission import NotEnoughPermission
+from ai.backend.manager.errors.repository import EntityNotFoundError
+from ai.backend.testutils.typed_scenario import (
+    TypedScenario,
+    at,
+    checked,
+    exactly,
+    ignored,
+)
+
+SCENARIOS: list[DomainScenario] = [
+    TypedScenario.ok(
+        "superadmin-edits-a-description-and-the-rest-is-left-alone",
+        given=[already_there("editable")],
+        when=admin_update("editable", UpdateDomainInput(description="edited")),
+        then=exactly(
+            DomainPayload(
+                domain=DomainNode(
+                    id=DomainID(UUID(int=0)),
+                    basic_info=DomainBasicInfo(
+                        name="editable", description="edited", integration_name=None
+                    ),
+                    registry=DomainRegistryInfo(allowed_docker_registries=[]),
+                    lifecycle=DomainLifecycleInfo(
+                        is_active=True,
+                        is_default=False,
+                        created_at=datetime.now(UTC),
+                        modified_at=datetime.now(UTC),
+                    ),
+                )
+            ),
+            where=(
+                ignored(lambda p: p.domain.id, "generated by the database"),
+                checked(lambda p: p.domain.lifecycle.created_at, WITHIN_THE_RUN, "just written"),
+                checked(lambda p: p.domain.lifecycle.modified_at, WITHIN_THE_RUN, "just edited"),
+            ),
+        ),
+    ),
+    TypedScenario.ok(
+        "retiring-a-domain-is-an-edit-of-its-active-flag",
+        given=[already_there("to-retire")],
+        when=admin_update("to-retire", UpdateDomainInput(is_active=False)),
+        then=at(lambda p: p.domain.lifecycle.is_active, False),
+    ),
+    TypedScenario.error(
+        "editing-a-name-nothing-answers-to-is-not-found",
+        when=admin_update("no-such-domain", UpdateDomainInput(description="x")),
+        then=EntityNotFoundError,
+    ),
+    TypedScenario.error(
+        "a-member-may-not-edit-a-domain",
+        actor=MEMBER,
+        given=[already_there("member-cannot-edit")],
+        when=admin_update("member-cannot-edit", UpdateDomainInput(description="x")),
+        then=NotEnoughPermission,
+    ),
+    TypedScenario.error(
+        "the-domain-admin-may-not-edit-the-domain-either",
+        actor=DOMAIN_ADMIN,
+        given=[already_there("admin-cannot-edit")],
+        when=admin_update("admin-cannot-edit", UpdateDomainInput(description="x")),
+        then=NotEnoughPermission,
+    ),
+]
+
+
+@pytest.mark.parametrize("scenario", SCENARIOS, ids=lambda s: s.id)
+async def test_editing(scenario: DomainScenario, run: TypedRunner) -> None:
+    await run(scenario)
